[ADD] Ler Comanda Vendedor / Autorizacao por PDV

+++ negocios/src/stores/sincronizacao.js
@@ -1,20 +1,41 @@
 import { defineStore } from "pinia";
 import { api } from "boot/axios";
 import { db } from "boot/db";
+import { uid } from "quasar";
+import { Platform } from "quasar";
+import { Notify } from "quasar";
 
 export const sincronizacaoStore = defineStore("sincronizacao", {
-  persist: true,
+  persist: {
+    paths: ["ultimaSincronizacao", "pdv"],
+  },
 
   state: () => ({
-    resultadoPesquisa: [],
-    textoPesquisa: null,
-    dialogPesquisa: false,
-    ultimaSincronizacaoFormaPagamento: null,
-    ultimaSincronizacaoEstoqueLocal: null,
-    ultimaSincronizacaoNaturezaOperacao: null,
-    ultimaSincronizacaoPessoa: null,
-    ultimaSincronizacaoProduto: null,
+    ultimaSincronizacao: {
+      FormaPagamento: null,
+      EstoqueLocal: null,
+      NaturezaOperacao: null,
+      Pessoa: null,
+      Produto: null,
+    },
     labelSincronizacao: "",
+    pdv: {
+      // envia para api
+      id: null,
+      latitude: null,
+      longitude: null,
+      precisao: null,
+      plataforma: null,
+      navegador: null,
+      versaonavegador: null,
+      desktop: null,
+      // backend retorna
+      autorizado: false,
+      apelido: null,
+      codfilial: null,
+      filial: null,
+      codpdv: null,
+    },
     importacao: {
       totalRegistros: null,
       totalSincronizados: null,
@@ -28,7 +49,69 @@ export const sincronizacaoStore = defineStore("sincronizacao", {
   }),
 
   actions: {
+    async dispositivo() {
+      if (!this.pdv.id) {
+        this.pdv.id = uid();
+      }
+
+      try {
+        const pos = await new Promise((resolve, reject) => {
+          navigator.geolocation.getCurrentPosition(resolve, reject);
+        });
+        this.pdv.latitude = pos.coords.latitude;
+        this.pdv.longitude = pos.coords.longitude;
+        this.pdv.precisao = pos.coords.accuracy;
+      } catch (error) {
+        Notify.create({
+          type: "negative",
+          message: error.message,
+          timeout: 0, // 20 minutos
+          actions: [{ icon: "close", color: "white" }],
+        });
+        var audio = new Audio("erro.mp3");
+        audio.play();
+      }
+
+      const plat = Platform.is;
+      this.pdv.plataforma = plat.platform;
+      this.pdv.navegador = plat.name;
+      this.pdv.versaonavegador = plat.version;
+      this.pdv.desktop = plat.desktop ? 1 : 0;
+      const params = {
+        uuid: this.pdv.id,
+        latitude: this.pdv.latitude,
+        longitude: this.pdv.longitude,
+        precisao: this.pdv.precisao,
+        desktop: this.pdv.desktop,
+        navegador: this.pdv.navegador,
+        versaonavegador: this.pdv.versaonavegador,
+        plataforma: this.pdv.plataforma,
+      };
+      let { data } = await api.put("/api/v1/pdv/dispositivo", params);
+      this.pdv.autorizado = data.data.autorizado;
+      this.pdv.apelido = data.data.apelido;
+      this.pdv.codfilial = data.data.codfilial;
+      this.pdv.filial = data.data.filial;
+      this.pdv.codpdv = data.data.codpdv;
+    },
+
     async sincronizar() {
+      // verifica se PDV pode acessar API
+      await this.dispositivo();
+      if (!this.pdv.autorizado) {
+        Notify.create({
+          type: "negative",
+          message:
+            "PDV não autorizado! Solicite autorização para o TI! \n UUID: " +
+            this.pdv.id,
+          timeout: 0, // 20 minutos
+          actions: [{ icon: "close", color: "white" }],
+        });
+        var audio = new Audio("erro.mp3");
+        audio.play();
+        return;
+      }
+
       // mostra janela de progresso
       this.importacao.rodando = true;
 
@@ -64,7 +147,9 @@ export const sincronizacaoStore = defineStore("sincronizacao", {
 
       try {
         // busca registros na ApI
-        let { data } = await api.get("/api/v1/pdv/impressora");
+        let { data } = await api.get("/api/v1/pdv/impressora", {
+          params: { uuid: this.pdv.id },
+        });
 
         // insere dados no banco local indexeddb
         await db.impressora.bulkPut(data);
@@ -88,7 +173,9 @@ export const sincronizacaoStore = defineStore("sincronizacao", {
 
       try {
         // busca registros na ApI
-        let { data } = await api.get("/api/v1/pdv/forma-pagamento");
+        let { data } = await api.get("/api/v1/pdv/forma-pagamento", {
+          params: { uuid: this.pdv.id },
+        });
 
         // insere dados no banco local indexeddb
         await db.formaPagamento.bulkPut(data);
@@ -98,7 +185,7 @@ export const sincronizacaoStore = defineStore("sincronizacao", {
         db.formaPagamento.where("sincronizado").below(sincronizado).delete();
 
         //registra data de Sincronizacao
-        this.ultimaSincronizacaoFormaPagamento = sincronizado;
+        this.ultimaSincronizacao.FormaPagamento = sincronizado;
       } catch (error) {
         console.log(error);
         console.log("Impossível sincronizar Formas de Pagamento");
@@ -112,7 +199,9 @@ export const sincronizacaoStore = defineStore("sincronizacao", {
 
       try {
         // busca registros na ApI
-        let { data } = await api.get("/api/v1/pdv/estoque-local");
+        let { data } = await api.get("/api/v1/pdv/estoque-local", {
+          params: { uuid: this.pdv.id },
+        });
 
         // insere dados no banco local indexeddb
         await db.estoqueLocal.bulkPut(data);
@@ -122,7 +211,7 @@ export const sincronizacaoStore = defineStore("sincronizacao", {
         db.estoqueLocal.where("sincronizado").below(sincronizado).delete();
 
         //registra data de Sincronizacao
-        this.ultimaSincronizacaoEstoqueLocal = sincronizado;
+        this.ultimaSincronizacao.EstoqueLocal = sincronizado;
       } catch (error) {
         console.log(error);
         console.log("Impossível sincronizar Locais de Estoque");
@@ -136,7 +225,9 @@ export const sincronizacaoStore = defineStore("sincronizacao", {
 
       try {
         // busca registros na ApI
-        let { data } = await api.get("/api/v1/pdv/natureza-operacao");
+        let { data } = await api.get("/api/v1/pdv/natureza-operacao", {
+          params: { uuid: this.pdv.id },
+        });
 
         // insere dados no banco local indexeddb
         await db.naturezaOperacao.bulkPut(data);
@@ -146,7 +237,7 @@ export const sincronizacaoStore = defineStore("sincronizacao", {
         db.naturezaOperacao.where("sincronizado").below(sincronizado).delete();
 
         //registra data de Sincronizacao
-        this.ultimaSincronizacaoNaturezaOperacao = sincronizado;
+        this.ultimaSincronizacao.NaturezaOperacao = sincronizado;
       } catch (error) {
         console.log(error);
         console.log("Impossível sincronizar Natureza Operacao");
@@ -159,7 +250,9 @@ export const sincronizacaoStore = defineStore("sincronizacao", {
 
       // descobre o total de registros pra sincronizar
       try {
-        let { data } = await api.get("/api/v1/pdv/pessoa-count");
+        let { data } = await api.get("/api/v1/pdv/pessoa-count", {
+          params: { uuid: this.pdv.id },
+        });
         this.importacao.totalRegistros = data.count;
         this.importacao.limiteRequisicao = Math.round(
           this.importacao.totalRegistros / 10
@@ -177,6 +270,7 @@ export const sincronizacaoStore = defineStore("sincronizacao", {
         // busca dados na api
         var { data } = await api.get("/api/v1/pdv/pessoa", {
           params: {
+            uuid: this.pdv.id,
             codpessoa: codpessoa,
             limite: this.importacao.limiteRequisicao,
           },
@@ -214,8 +308,10 @@ export const sincronizacaoStore = defineStore("sincronizacao", {
       );
 
       // exclui registros que nao vieram na importacao
-      db.pessoa.where("sincronizado").below(sincronizado).delete();
-      this.ultimaSincronizacaoPessoa = sincronizado;
+      if (this.importacao.rodando) {
+        db.pessoa.where("sincronizado").below(sincronizado).delete();
+      }
+      this.ultimaSincronizacao.Pessoa = sincronizado;
     },
 
     async sincronizarProduto() {
@@ -224,7 +320,9 @@ export const sincronizacaoStore = defineStore("sincronizacao", {
 
       // descobre o total de registros pra sincronizar
       try {
-        let { data } = await api.get("/api/v1/pdv/produto-count");
+        let { data } = await api.get("/api/v1/pdv/produto-count", {
+          params: { uuid: this.pdv.id },
+        });
         this.importacao.totalRegistros = data.count;
         this.importacao.limiteRequisicao = Math.round(
           this.importacao.totalRegistros / 50
@@ -242,6 +340,7 @@ export const sincronizacaoStore = defineStore("sincronizacao", {
         // busca dados na api
         var { data } = await api.get("/api/v1/pdv/produto", {
           params: {
+            uuid: this.pdv.id,
             codprodutobarra: codprodutobarra,
             limite: this.importacao.limiteRequisicao,
           },
@@ -279,8 +378,10 @@ export const sincronizacaoStore = defineStore("sincronizacao", {
       );
 
       // exclui registros que nao vieram na importacao
-      db.produto.where("sincronizado").below(sincronizado).delete();
-      this.ultimaSincronizacaoProduto = sincronizado;
+      if (this.importacao.rodando) {
+        db.produto.where("sincronizado").below(sincronizado).delete();
+      }
+      this.ultimaSincronizacao.Produto = sincronizado;
     },
   },
 });
